+++ inst/realtest/ISOdatetime-all.R
@@ -1,5 +1,6 @@
 E(
     as.integer(ISOdate(1970, 1, 1, tz="UTC")),
     0L,
-    bad=12L*60L*60L
+    bad=12L*60L*60L,
+    value_comparer=all.equal  # storage mode may be 'double' or 'integer'
 )