+++ tests/realtest.R
@@ -1,5 +1,5 @@
 # This file is part of the 'stringx' project.
-# Copyleft (c) 2021-2024, Marek Gagolewski <https://www.gagolewski.com>
+# Copyleft (c) 2021-2025, Marek Gagolewski <https://www.gagolewski.com>
 
 # Runs all unit tests for the package